backend/v1/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-filter==24.3
django-mysql==4.15.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.2
Markdown==3.7
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
PyMySQL==1.1.1
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
setuptools==76.1.0
simplejwt==2.0.1
six==1.17.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
